feat!: remove Attachment type in favour of MediaAttachment
