+++ Sources/TootSDK/Models/MediaAttachment.swift
@@ -3,6 +3,7 @@
 
 import Foundation
 
+/// Represents a file or media attachment that can be added to a status.
 public struct MediaAttachment: Codable, Hashable, Sendable {
     
     public init(id: String, type: AttachmentType, url: String? = nil, remoteUrl: String? = nil, previewUrl: String? = nil, meta: AttachmentMeta? = nil, description: String? = nil, blurhash: String? = nil) {
@@ -32,4 +33,60 @@ public struct MediaAttachment: Codable, Hashable, Sendable {
     public var description: String?
     /// A hash computed by the BlurHash algorithm, for generating colorful preview thumbnails when media has not been downloaded yet.
     public var blurhash: String?
+}
+
+public enum AttachmentType: String, Codable, Hashable, Sendable {
+    ///  Static image
+    case image
+    /// Video clip
+    case video
+    /// Looping, soundless animation
+    case gifv
+    /// Audio track
+    case audio
+    /// unsupported or unrecognized file type
+    case unknown
+}
+
+public struct AttachmentMeta: Codable, Hashable, Sendable {
+
+    public var original: AttachmentMetaInfo?
+    public var small: AttachmentMetaInfo?
+    public var focus: AttachmentMetaFocus?
+}
+
+public struct AttachmentMetaInfo: Codable, Hashable, Sendable {
+    public var width: Int?
+    public var height: Int?
+    public var size: String?
+    public var aspect: Double?
+    public var frameRate: String?
+    public var duration: Double?
+    public var bitrate: Int?
+}
+
+public extension MediaAttachment {
+    var aspectRatio: Double? {
+        if
+            let info = meta?.original,
+            let width = info.width,
+            let height = info.height,
+            width != 0,
+            height != 0 {
+            let aspectRatio = Double(width) / Double(height)
+            
+            return aspectRatio.isNaN ? nil : aspectRatio
+        }
+        
+        return nil
+    }
+}
+
+public struct AttachmentMetaFocus: Codable, Hashable, Sendable {
+    public var x: Double
+    public var y: Double
+}
+
+public extension AttachmentMetaFocus {
+    static let `default` = Self(x: 0, y: 0)
 }
